Add BrowserRouter, components and User Interface
Develop front end of the project to have a shell to work with

Route can still not be imported from react-router-dom, hence
only the path / is rendered. Cleaned import structure of all
components. Material-UI was chosen for the UI-design as it
is very responsive. Build welcome and dj_or_guest path.